added more logic, questions, and statements.
Put query commands on top of statement commands so they are considered first to avoid defining people named "who". Added a lot of relationship definitions, questions, and statements. Needs more precise logic in add statements. replace relationship with a predicate that defines all possible ways to make a relationship being added as false to ensure intelligence of chatbot is correct.